requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.4
dill==0.3.9
exceptiongroup==1.2.2
fastapi==0.115.4
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
playwright==1.48.0
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
pylint==3.3.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-pylint==0.21.0
requests==2.32.3
sniffio==1.3.1
starlette==0.41.2
tomli==2.0.2
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
fastapi
sqlalchemy
pydantic[email]
pytest
psycopg2-binary
passlib[bcrypt]
PyJWT==2.8.0

